+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-## [0.28.05] - 2021-12-11
+## [0.28.06] - 2021-12-12
 ### Changed
  - wc command.
    - Fixed the bug that did not count the number of rows of data passed from PIPE correctly.
@@ -24,6 +24,8 @@ All notable changes to this project will be documented in this file. The format 
  - mkfifo command.
    - Changed to continue processing even if multiple named pipes fail while creating.
    - Changed to print the path of the file that failed to be created on error.
+ - dos2unix command.
+   - Changed to print the message being converted.
 ## [0.28.00] - 2021-12-08
 ### Added
  - sddf command. Search & Delete Duplicated Files.